refactor(cli): setup Phase B (cmdSetup+cmdInit) + review → commands/
Twelfth cut of the cli.ts god-module split. With install/login/check/design
all extracted, the orchestrators they drive are now safe to move too.

- Move cmdSetup + cmdInit into src/commands/setup.ts (joining Phase A's
  cmdDoctor + cmdAdd). Their cluster-private helpers move with them and stay
  unexported: runConfiguredCommand, generateConfigFile, offerContextLock,
  offerPosture, offerFirstInstallPrescan.
- New src/commands/review.ts — cmdReview (check + design + standards meta-run).
- cli.ts imports cmdSetup/cmdInit/cmdReview back; COMMANDS table unchanged.
  Pruned now-unused imports.

Verification: tsc clean, eslint 0 errors, full suite 2580 passing, prettier
clean, runtime dispatch smoke-tested (setup/init/review).

cli.ts: 2910 → 2145 lines (−765; ~66% smaller than the 6330-line original).
